refactor multiprocessing code to greatly simplify
